Util refactor
