+++ src/utils/deploy/util.js
@@ -1,0 +1,13 @@
+const path = require('path')
+
+exports.normalizePath = relname => {
+  return (
+    '/' +
+    relname
+      .split(path.sep)
+      .map(segment => {
+        return encodeURIComponent(segment)
+      })
+      .join('/')
+  )
+}
